Переделка из-за вопросов
