+++ package.json
@@ -20,5 +20,8 @@
   },
   "engines": {
     "node": "10"
+  },
+  "dependencies": {
+    "dayjs": "1.9.5"
   }
 }
