+++ src/utils/tools.js
@@ -1,8 +1,8 @@
 import dayjs from 'dayjs';
 
-import { TRIP_POINT_TYPES, CITIES, DESCRIPTION, OFFERS } from './const';
+import {TRIP_POINT_TYPES, CITIES, DESCRIPTION, OFFERS} from '../const';
 
-export function getId(t=21) {
+export function getId(t = 21) {
   let e="",r=crypto.getRandomValues(new Uint8Array(t));for(;t--;){let n=63&r[t];e+=n<36?n.toString(36):n<62?(n-26).toString(36).toUpperCase():n<63?"_":"-"}return e;
 }
 
@@ -11,15 +11,15 @@ export function getRandomInteger(a = 0, b = 1) {
   const upper = Math.floor(Math.max(a, b));
 
   return Math.floor(lower + Math.random() * (upper - lower + 1));
-};
+}
 
 export function getRandomType() {
   return TRIP_POINT_TYPES[getRandomInteger(0, TRIP_POINT_TYPES.length - 1)];
-};
+}
 
 export function getRandomCity() {
   return CITIES[getRandomInteger(0, CITIES.length - 1)];
-};
+}
 
 export function getRandomDescription() {
   const description = [];
@@ -30,21 +30,21 @@ export function getRandomDescription() {
     description.push(sentence);
   }
 
-  return description.join(' ');
-};
+  return description.join(` `);
+}
 
 export function getRandomPrice(maxPrice = 100) {
   return getRandomInteger(1, maxPrice) * 10;
-};
+}
 
 export function getRandomImages() {
   const imagesNumber = getRandomInteger(1, 5);
-  const images = []
+  const images = [];
   for (let i = 0; i < imagesNumber; i++) {
     images.push(`http://picsum.photos/248/152?r=${Math.random()}`);
   }
   return images;
-};
+}
 
 export function getRandomDates() {
   const offset = getRandomInteger(1, 10);
@@ -52,7 +52,7 @@ export function getRandomDates() {
   const maxDaysGap = 7;
   const daysGap = getRandomInteger(-maxDaysGap, maxDaysGap);
   const startDate = dayjs().add(daysGap, `day`);
-  const endDate = startDate.add(offset, 'hour');
+  const endDate = startDate.add(offset, `hour`);
 
   return [startDate.toDate(), endDate.toDate()];
 }
