+++ src/view/cost.js
@@ -1,4 +1,4 @@
-import {createElement} from '../utils/utils';
+import {createElement} from '../utils';
 
 function createCost(cost) {
   return `<p class="trip-info__cost">
