+++ src/view/events-sort-form.js
@@ -1,4 +1,4 @@
-import {createElement} from '../utils/utils';
+import {createElement} from '../utils';
 
 function createTripEventsSortForm() {
   return `
